chore: rename set-up-service-worker
