+++ pwa/src/index.js
@@ -1,4 +1,4 @@
-export { setUpServiceWorker } from './service-worker/service-worker.js'
+export { setUpServiceWorker } from './service-worker/set-up-service-worker.js'
 export { OfflineInterface } from './offline-interface/offline-interface.js'
 export {
     checkForUpdates,
