+++ pwa/src/service-worker/set-up-service-worker.js
